requirements.txt
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
pydantic==2.5.2
typing-extensions==4.8.0
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
datasets==2.16.1
openai==1.6.1
langchain==0.1.0
langchain-google-genai==0.0.6 